fixed a small issue

+++ 0x1E-search_algorithms/1-binary.c
@@ -15,9 +15,9 @@
  *              ends or value found
  */
 
-int binary_search(int *array, int size, int value)
+int binary_search(int *array, size_t size, int value)
 {
-    int left = 0, right = size - 1, mid;
+    size_t left = 0, right = size - 1, mid;
 
     if (array == NULL)
     {
